requirements.txt
Biopython
joblib
tqdm
networkx
numpy
pandas
pysam
scipy
matplotlib